Refactor testing cases (#189)
* refactor: clear up testing cases;

* refactor: refactor code in Dataset classes for models;

* refactor: adjust testing workflows according to refactored test cases;

* fix: turn missing_mask into torch.float;

* fix: error in BTTF testing case;

* feat: using pip to manage dependencies in CI testing workflow, and using conda in Daily testing workflow;
