+++ .github/workflows/testing_ci.yml
@@ -15,43 +15,60 @@ jobs:
         runs-on: ${{ matrix.os }}
         defaults:
             run:
-                shell: bash -l {0}
+                shell: bash {0}
         strategy:
             fail-fast: false
             matrix:
                 os: [ubuntu-latest, windows-latest, macOS-latest]
-                python-version: ["3.7", "3.8", "3.9", "3.10"]
+                python-version: ["3.7", "3.10"]
+                torch-version: ["1.13.1"]
 
         steps:
             - name: Check out the repo code
               uses: actions/checkout@v3
 
-            - name: Set up Conda
-              uses: conda-incubator/setup-miniconda@v2
+            - name: Determine the Python version
+              uses: haya14busa/action-cond@v1
+              id: condval
               with:
-                  activate-environment: pypots-test
-                  python-version: ${{ matrix.python-version }}
-                  environment-file: tests/environment_for_conda_test.yml
-                  auto-activate-base: false
+                cond: ${{ matrix.python-version == 3.7 && matrix.os == 'macOS-latest' }}
+                # Note: the latest 3.7 subversion 3.7.17 for MacOS has "ModuleNotFoundError: No module named '_bz2'"
+                if_true: "3.7.16"
+                if_false: ${{ matrix.python-version }}
+
+            - name: Set up Python
+              uses: actions/setup-python@v4
+              with:
+                python-version: ${{ steps.condval.outputs.value }}
+                check-latest: true
+                cache: pip
+                cache-dependency-path: |
+                  setup.cfg
+
+            - name: Install PyTorch ${{ matrix.torch-version }}+cpu
+              # we have to install torch in advance because torch_sparse needs it for compilation,
+              # refer to https://github.com/rusty1s/pytorch_sparse/issues/156#issuecomment-1304869772 for details
+              run: |
+                  which python
+                  which pip
+                  python -m pip install --upgrade pip
+                  pip install torch==${{ matrix.torch-version }} -f https://download.pytorch.org/whl/cpu
+                  python -c "import torch; print('PyTorch:', torch.__version__)"
+
+            - name: Install other dependencies
+              run: |
+                  pip install pypots
+                  pip install torch-geometric torch-scatter torch-sparse -f "https://data.pyg.org/whl/torch-${{ matrix.torch-version }}+cpu.html"
+                  pip install -e ".[dev]"
 
             - name: Fetch the test environment details
               run: |
                   which python
-                  conda info
-                  conda list
+                  pip list
 
             - name: Test with pytest
               run: |
-                  # run tests separately here due to Segmentation Fault in test_clustering when run all in
-                  # one command with `pytest` on MacOS. Bugs not caught, so this is a trade-off to avoid SF.
-                  python -m pytest -rA tests/test_classification.py -n auto --cov=pypots --dist=loadgroup --cov-config=.coveragerc
-                  python -m pytest -rA tests/test_imputation.py -n auto --cov=pypots --cov-append --dist=loadgroup --cov-config=.coveragerc
-                  python -m pytest -rA tests/test_clustering.py -n auto --cov=pypots --cov-append --dist=loadgroup --cov-config=.coveragerc
-                  python -m pytest -rA tests/test_forecasting.py -n auto --cov=pypots --cov-append --dist=loadgroup --cov-config=.coveragerc
-                  python -m pytest -rA tests/test_optim.py -n auto --cov=pypots --cov-append --dist=loadgroup --cov-config=.coveragerc
-                  python -m pytest -rA tests/test_data.py -n auto --cov=pypots --cov-append --dist=loadgroup --cov-config=.coveragerc
-                  python -m pytest -rA tests/test_utils.py -n auto --cov=pypots --cov-append --dist=loadgroup --cov-config=.coveragerc
-                  python -m pytest -rA tests/test_cli.py -n auto --cov=pypots --cov-append --dist=loadgroup --cov-config=.coveragerc
+                  coverage run --source=pypots -m pytest -rA tests/*/*
 
             - name: Generate the LCOV report
               run: |
@@ -61,4 +78,4 @@ jobs:
               uses: coverallsapp/github-action@master
               with:
                   github-token: ${{ secrets.GITHUB_TOKEN }}
-                  path-to-lcov: 'coverage.lcov'
+                  path-to-lcov: "coverage.lcov"
