+++ docs/pypots.forecasting.rst
@@ -1,11 +1,31 @@
 pypots.forecasting package
 ==========================
 
+Subpackages
+-----------
 
-pypots.forecasting.bttf module
+.. toctree::
+   :maxdepth: 4
+
+   pypots.forecasting.bttf
+   pypots.forecasting.template
+
+Submodules
+----------
+
+pypots.forecasting.base module
 ------------------------------
 
-.. automodule:: pypots.forecasting.bttf
+.. automodule:: pypots.forecasting.base
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :inherited-members:
+
+Module contents
+---------------
+
+.. automodule:: pypots.forecasting
    :members:
    :undoc-members:
    :show-inheritance:
