+++ pypots/classification/template/data.py
