+++ pypots/forecasting/template/data.py
