+++ tests/cli/config.py
@@ -1,0 +1,11 @@
+"""
+Test configs for CLI tools.
+"""
+
+# Created by Wenjie Du <wenjay.du@gmail.com>
+# License: GLP-v3
+
+import os
+
+
+PROJECT_ROOT_DIR = os.path.abspath(os.path.join(os.path.abspath(__file__), "../../.."))
