+++ tests/utils/logging.py
@@ -1,5 +1,5 @@
 """
-Test cases for the functions and classes in package `pypots.utils`.
+Test cases for the functions and classes in package `pypots.utils.logging`.
 """
 
 # Created by Wenjie Du <wenjay.du@gmail.com>
@@ -9,10 +9,7 @@ import os
 import shutil
 import unittest
 
-import torch
-
 from pypots.utils.logging import Logger
-from pypots.utils.random import set_random_seed
 
 
 class TestLogging(unittest.TestCase):
@@ -47,26 +44,6 @@ class TestLogging(unittest.TestCase):
         self.logger_creator.set_saving_path("test_log", "testing.log")
         assert os.path.exists("test_log/testing.log")
         shutil.rmtree("test_log", ignore_errors=True)
-
-
-class TestRandom(unittest.TestCase):
-    def test_set_random_seed(self):
-        random_state1 = torch.get_rng_state()
-        torch.rand(
-            1, 3
-        )  # randomly generate something, the random state will be reset, so two states should be varying
-        random_state2 = torch.get_rng_state()
-        assert not torch.equal(
-            random_state1, random_state2
-        ), "The random seed hasn't set, so two random states should be different."
-
-        set_random_seed(26)
-        random_state1 = torch.get_rng_state()
-        set_random_seed(26)
-        random_state2 = torch.get_rng_state()
-        assert torch.equal(
-            random_state1, random_state2
-        ), "The random seed has been set, two random states are not the same."
 
 
 if __name__ == "__main__":
